scripts/pyproject.toml
[tool.poetry]
name = "geoparquet-scripts"
version = "0.1.0"
description = ""
authors = []
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
pyarrow = "^7.0.0"
geopandas = "^0.10.2"
pygeos = "^0.12.0"
pandas = "^1.4.2"
click = "^8.1.2"
google-cloud-bigquery = "^3.2.0"
db-dtypes = "^1.0.2"

[tool.poetry.dev-dependencies]
ipykernel = "^6.13.0"
isort = "^5.10.1"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
